Add custom method to rails_helper
I wanted to experiment with adding custom methods
to rails_helper like Sean did, which we emulated in dinner
dash.  At first this did not work because I tried using the
method in the model test, and since it uses "page" the model test
didn't know what "page" was because it's inherited from
Capybara, which is in feature tests.  So, I moved it to
a feature test and it works.