Add thumbs down test
I added a test for thumbs down but
it incorrectly passed because the word "Swill"
appeared in the error text.

I then wanted to specify that the status code is 200
and had to look up the syntax for that.  Verifying the
status code before you verify something is on the page
is easier than using id's or scopes.