Grab form data
I want to console log the form data, so
I made a function formData for this which
returns an object with keys idea_title,
idea_body.  The first time I called this,
I only used formData, so it returned the
function itself, to get the actual function
to be called I had to use formData().